add handling for item card rolls with no .card-buttons element
When using the quick roll functionality from Ready Set Roll, there is no .card-buttons element.  Thus, even if the user has enabled the "Add Chat Buttons" option, the buttons will not be rendered.  This change checks for the presence of the .card-buttons element, and inserts a blank element prior to adding the "Add Convenient Effect" button if there is no .card-buttons element.